refactor(analytic-logs): Replace ProviderVolumeCounts with WorkflowActivityCounts for provider volume data retrieval
- Removed ProviderVolumeCounts table and repository.
- Updated SQL migrations to reflect the removal of provider volume counts.
- Adjusted use cases and shared module to utilize WorkflowActivityCountsRepository for provider volume data.
- Added provider_id to relevant schemas and queries for consistency.
