+++ libs/application-generic/src/services/analytic-logs/index.ts
@@ -2,7 +2,6 @@ export { createClient as createClickHouseClient } from '@clickhouse/client';
 export * from './clickhouse.service';
 export * from './delivery-trend-counts';
 export * from './log.repository';
-export * from './provider-volume-counts';
 export * from './request-log';
 export { StepRun, StepRunFinalStatus, StepRunNonFinalStatus, StepRunRepository, StepRunStatus } from './step-run';
 export { EventType, mapEventTypeToTitle, Trace, TraceLogRepository, TraceStatus, traceLogSchema } from './trace-log';
