--- libs/application-generic/src/services/analytic-logs/provider-volume-counts/index.ts
@@ -1,2 +1,0 @@
-export * from './provider-volume-counts.repository';
-export * from './provider-volume-counts.schema';
